Add support for PREAPPROVED_GITHUB_ACCOUNTS
This feature allows certain Github usernames to be pre-approved, i.e. to not
require CLAs present in the database for any email. It's especially useful for
bots which create commits (like dependabot) to not trigger the CLA warning.

This is a minimal implementation for now to hotfix an issue in CPython
workflow. Ideally it should be implemented as a database object type.